Added fix for Paging data.

+++ Epi Info Web Survey/Epi.Web.EF/EntitySurveyInfoDao.cs
@@ -44,21 +44,23 @@ namespace Epi.Web.EF
             }
 
 
-            result.Sort(CompareByDateCreated);
+            
 
             // remove the items to skip
             // remove the items after the page size
             if (PageNumber > 0 && PageSize > 0)
             {
+                result.Sort(CompareByDateCreated);
                 // remove the items to skip
                 if (PageNumber * PageSize - PageSize > 0)
                 {
                     result.RemoveRange(0, PageSize);
                 }
 
-                // remove the items after the page size
-                result.RemoveRange(PageNumber * PageSize, result.Count - PageNumber * PageSize);
-
+                if (PageNumber * PageSize < result.Count)
+                {
+                    result.RemoveRange(PageNumber * PageSize, result.Count - PageNumber * PageSize);
+                }
             }
 
             return result;
@@ -111,19 +113,23 @@ namespace Epi.Web.EF
             }
 
             result = Mapper.Map(responseList);
-            result.Sort(CompareByDateCreated);
+            
 
             // remove the items to skip
             // remove the items after the page size
             if (PageNumber > 0 && PageSize > 0)
             {
+                result.Sort(CompareByDateCreated);
                 // remove the items to skip
                 if (PageNumber * PageSize - PageSize > 0)
                 {
                     result.RemoveRange(0, PageSize);
                 }
 
-                result.RemoveRange(PageNumber * PageSize, result.Count - PageNumber * PageSize);
+                if (PageNumber * PageSize < result.Count)
+                {
+                    result.RemoveRange(PageNumber * PageSize, result.Count - PageNumber * PageSize);
+                }
             }
             
             return result;
